+++ routes/subTodoRouter.js
@@ -1,14 +1,17 @@
 const express = require('express');
-const subTodo = require('../models/subTodo');
+const { subTodo } = require('../models/subTodo');
 const authenticate = require('../authenticate');
+const cors = require('./cors');
 
 const subTodoRouter = express.Router();
 
 subTodoRouter
     .route('/')
-    .get(authenticate.verifyUser, (req, res, next) => {
+    .options(cors.corsWithOptions, (req, res) => res.sendStatus(200))
+
+    .get(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
         subTodo
-            .find()
+            .find({ userId: req.user._id })
             .then((subTodos) => {
                 res.statusCode = 200;
                 res.setHeader('Content-Type', 'application/json');
@@ -17,26 +20,26 @@ subTodoRouter
             .catch((err) => next(err));
     })
 
-    .post(authenticate.verifyUser, (req, res, next) => {
-        subTodo
-            .create(req.body)
-            .then((SubTodo) => {
-                console.log('SubTodo has been created', SubTodo);
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(SubTodo);
+    .post(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
+        const newSubTodo = new subTodo({ ...req.body, userId: req.user._id });
+        newSubTodo
+            .save()
+            .then((savedSubTodo) => {
+                res.status(201).json(savedSubTodo);
             })
-            .catch((err) => next(err));
+            .catch((err) => {
+                next(err);
+            });
     })
 
-    .put(authenticate.verifyUser, (req, res) => {
+    .put(cors.corsWithOptions, authenticate.verifyUser, (req, res) => {
         res.statusCode = 403;
         res.end('PUT opration not supported on /subTodos');
     })
 
-    .delete(authenticate.verifyUser, (req, res, next) => {
+    .delete(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
         subTodo
-            .deleteMany()
+            .deleteMany({ userId: req.user._id })
             .then((response) => {
                 res.statusCode = 200;
                 res.setHeader('Content-Type', 'application/json');
@@ -47,7 +50,9 @@ subTodoRouter
 
 subTodoRouter
     .route('/:subTodoId')
-    .get(authenticate.verifyUser, (req, res, next) => {
+    .options(cors.corsWithOptions, (req, res) => res.sendStatus(200))
+
+    .get(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
         subTodo
             .findById(req.params.subTodoId)
             .then((todo) => {
@@ -58,14 +63,14 @@ subTodoRouter
             .catch((err) => next(err));
     })
 
-    .post(authenticate.verifyUser, (req, res) => {
+    .post(cors.corsWithOptions, authenticate.verifyUser, (req, res) => {
         res.status = 403;
         res.end(
             `POST operation not supported on /todos/${req.params.subTodoId}`
         );
     })
 
-    .put(authenticate.verifyUser, (req, res, next) => {
+    .put(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
         subTodo
             .findByIdAndUpdate(
                 req.params.subTodoId,
@@ -80,7 +85,7 @@ subTodoRouter
             .catch((err) => next(err));
     })
 
-    .delete(authenticate.verifyUser, (req, res, next) => {
+    .delete(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
         subTodo
             .findByIdAndDelete(req.params.subTodoId)
             .then((response) => {
